Fix various code style issues
Most importantly, don't use module names that are the same as the contained class names,
as this can lead to errors where the import order matters, as classes are confused with their modules.
(For example, this was the case if you sorted imports of hcs/imagexpress/__init__.py before this commit.)

We now collect several classes into single (sub-)modules.

Usages of `xml`/`lxml` where replaced by alternatives from `defusedxml`.

This commit also fixes various other errors reported by `hatch fmt` (i.e. ruff).
