+++ .gitignore
@@ -91,3 +91,4 @@ mobie-example-project/
 
 # scratchpad
 scratchpad/
+.pixi
