+++ src/faim_ipa/io/metadata.py
