+++ src/faim_ipa/io/metaseries.py
