+++ src/faim_ipa/stitching/bounding_box.py
