+++ src/faim_ipa/utils.py
@@ -1,7 +1,7 @@
 import logging
+import os.path
 import pathlib
 from datetime import datetime
-import os.path
 
 import pydantic
 from pydantic import BaseModel
@@ -19,40 +19,40 @@ def wavelength_to_rgb(wavelength, gamma=0.8):
     """
 
     wavelength = float(wavelength)
-    if 380 <= wavelength <= 440:
+    if 380 <= wavelength <= 440:  # noqa: PLR2004
         attenuation = 0.3 + 0.7 * (wavelength - 380) / (440 - 380)
-        R = ((-(wavelength - 440) / (440 - 380)) * attenuation) ** gamma
-        G = 0.0
-        B = (1.0 * attenuation) ** gamma
-    elif 440 <= wavelength <= 490:
-        R = 0.0
-        G = ((wavelength - 440) / (490 - 440)) ** gamma
-        B = 1.0
-    elif 490 <= wavelength <= 510:
-        R = 0.0
-        G = 1.0
-        B = (-(wavelength - 510) / (510 - 490)) ** gamma
-    elif 510 <= wavelength <= 580:
-        R = ((wavelength - 510) / (580 - 510)) ** gamma
-        G = 1.0
-        B = 0.0
-    elif 580 <= wavelength <= 645:
-        R = 1.0
-        G = (-(wavelength - 645) / (645 - 580)) ** gamma
-        B = 0.0
-    elif 645 <= wavelength <= 750:
+        r = ((-(wavelength - 440) / (440 - 380)) * attenuation) ** gamma
+        g = 0.0
+        b = (1.0 * attenuation) ** gamma
+    elif 440 <= wavelength <= 490:  # noqa: PLR2004
+        r = 0.0
+        g = ((wavelength - 440) / (490 - 440)) ** gamma
+        b = 1.0
+    elif 490 <= wavelength <= 510:  # noqa: PLR2004
+        r = 0.0
+        g = 1.0
+        b = (-(wavelength - 510) / (510 - 490)) ** gamma
+    elif 510 <= wavelength <= 580:  # noqa: PLR2004
+        r = ((wavelength - 510) / (580 - 510)) ** gamma
+        g = 1.0
+        b = 0.0
+    elif 580 <= wavelength <= 645:  # noqa: PLR2004
+        r = 1.0
+        g = (-(wavelength - 645) / (645 - 580)) ** gamma
+        b = 0.0
+    elif 645 <= wavelength <= 750:  # noqa: PLR2004
         attenuation = 0.3 + 0.7 * (750 - wavelength) / (750 - 645)
-        R = (1.0 * attenuation) ** gamma
-        G = 0.0
-        B = 0.0
+        r = (1.0 * attenuation) ** gamma
+        g = 0.0
+        b = 0.0
     else:
-        R = 0.0
-        G = 0.0
-        B = 0.0
-    R *= 255
-    G *= 255
-    B *= 255
-    return int(R), int(G), int(B)
+        r = 0.0
+        g = 0.0
+        b = 0.0
+    r *= 255
+    g *= 255
+    b *= 255
+    return int(r), int(g), int(b)
 
 
 def rgb_to_hex(r, g, b):
@@ -70,7 +70,7 @@ def create_logger(name: str) -> logging.Logger:
     name
         Name of the logger instance.
     """
-    logger = logging.Logger(name.capitalize())
+    logger = logging.getLogger(name.capitalize())
     now = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
     handler = logging.FileHandler(f"{now}-{name}.log")
     handler.setLevel(logging.INFO)
@@ -148,10 +148,11 @@ class IPAConfig(BaseModel):
         The paths are assumed to be relative to a git-root directory somewhere
         in the parent directories of the class implementing `IPAConfig`.
         """
-        if pydantic.__version__.startswith("2"):
-            fields = self.model_fields_set
-        else:
-            fields = self.__fields_set__
+        fields = (
+            self.model_fields_set
+            if pydantic.__version__.startswith("2")
+            else self.__fields_set__
+        )
 
         for f in fields:
             attr = getattr(self, f)
@@ -166,10 +167,11 @@ class IPAConfig(BaseModel):
         somewhere in the parent directories of the class implementing
         `IPAConfig`.
         """
-        if pydantic.__version__.startswith("2"):
-            fields = self.model_fields_set
-        else:
-            fields = self.__fields_set__
+        fields = (
+            self.model_fields_set
+            if pydantic.__version__.startswith("2")
+            else self.__fields_set__
+        )
 
         for f in fields:
             attr = getattr(self, f)
