+++ src/faim_ipa/visiview/__init__.py
@@ -1,0 +1,11 @@
+from faim_ipa.visiview.acquisition import (
+    RegionAcquisitionOME,
+    RegionAcquisitionSTK,
+    StackedTile,
+)
+
+__all__ = [
+    "RegionAcquisitionOME",
+    "RegionAcquisitionSTK",
+    "StackedTile",
+]
