+++ tests/hcs/cellvoyager/test_CellVoyagerWellAcquisition.py
@@ -6,10 +6,10 @@ import pandas as pd
 import pytest
 
 from faim_ipa.hcs.acquisition import TileAlignmentOptions
-from faim_ipa.hcs.cellvoyager.CellVoyagerWellAcquisition import (
+from faim_ipa.hcs.cellvoyager import (
     CellVoyagerWellAcquisition,
+    StackedTile,
 )
-from faim_ipa.hcs.cellvoyager.StackedTile import StackedTile
 
 
 @pytest.fixture
